Task2/submit_task2/e2e/mc_ocr_rivf2020/requirements.txt
easydict 
imutils 
pandas 
Pillow 
matplotlib 
numpy 
scipy 
requests 
Cython
scikit-learn
tqdm 
recognizers_text 
recognizers_date_time 
pyyaml 
lmdb
unidecode
Levenshtein
shapely
flask-restful
Flask-Cors
